Extract Slider component to SliderEx
